+++ pkg/log/config.go
@@ -1,5 +1,7 @@
 package log
 
+import "time"
+
 // Config  log config
 type Config struct {
 	Development       bool
@@ -14,4 +16,5 @@ type Config struct {
 	LogFormatText     bool
 	LogRollingPolicy  string
 	LogBackupCount    uint
+	FlushInterval     time.Duration // default is 30s, recommend is dev or test is 1s, prod is 1m
 }