+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## Changelog
+# Changelog
+
+## v1.10.0
+
+- feat: support async flush log to dish
+- chore: using gorm offical plguin for tracing and metrics
 
 ## v1.9.0